move dependency injection into domain
